USDZ Loader - fix index array type
- Current `Uint16Array` type might cause spaghetti-like effect when opening some USDZ models
- Changed to `Uint32Array` type

+++ examples/jsm/loaders/USDZLoader.js
@@ -368,7 +368,7 @@ class USDZLoader extends Loader {
 			if ( 'int[] faceVertexIndices' in data ) {
 
 				const indices = JSON.parse( data[ 'int[] faceVertexIndices' ] );
-				geometry.setIndex( new BufferAttribute( new Uint16Array( indices ), 1 ) );
+				geometry.setIndex( new BufferAttribute( new Uint32Array( indices ), 1 ) );
 
 			}
